fix create temporary directory

+++ src/agat/agat_sp_filter_feature_from_kill_list/test.sh
@@ -8,11 +8,11 @@ set -eo pipefail
 test_dir="${meta_resources_dir}/test_data"
 
 # create temporary directory and clean up on exit
-TMPDIR=$(mktemp -d --tmpdir "$meta_temp_dir")
+TMPDIR=$(mktemp -d "$meta_temp_dir/$meta_functionality_name-XXXXXX")
 function clean_up {
- rm -rf "$TMPDIR"
+ [[ -d "$TMPDIR" ]] && rm -rf "$TMPDIR"
 }
-trap clean_up EXIT
+#trap clean_up EXIT
 
 echo "> Run $meta_name with test data"
 "$meta_executable" \
